NF: moving CheckMedia where it's used
